Partial script update
need to figure out how to use 2 cargo.toml
