+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 node_rust/target/
+node_rust/.DS_Store
 dist
-.DS_Store
