+++ node_rust/Cargo.toml
@@ -8,7 +8,7 @@ crate-type = ["cdylib", "rlib"]
 
 [[bin]]
 name = "main"
-path = "src/main.rs"
+path = "src/lib.rs"
 
 [dependencies]
 serde = { version = "1.0.104", features = ["derive"] }
@@ -18,3 +18,4 @@ wasm-bindgen = "0.2.63"
 [dependencies.web-sys]
 version = "0.3"
 features = [ "console" ]
+
