+++ node_rust/src/lib.rs
@@ -1,12 +1,15 @@
-extern crate wasm_bindgen;
-use web_sys::console;
-
-use wasm_bindgen::prelude::*;
-
+extern crate serde_json;
 mod rules;
+use std::fs;
+use std::io::Read;
 
-#[wasm_bindgen]
-pub fn main(data_string: String) -> String {
-    console::log_1(&"Greetins from Rustlandia!".into());
-    return rules::run(data_string);
+fn main() {
+    let mut file = fs::File::open("input.json").unwrap();
+    let mut data_string = String::new();
+
+    file.read_to_string(&mut data_string).unwrap();
+
+    let processed_data_string = rules::run(data_string);
+    println!("{}", processed_data_string);
+    fs::write("output.json", processed_data_string).expect("Unable to write file");
 }
